pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillctl"
version = "0.1.0"
description = "The package manager for agent skills. Search, install, lint SKILL.md files. Zero-config CLI."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Deval"},
]
keywords = ["cli", "agent", "skills", "claude", "codex", "package-manager"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Software Distribution",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/dvlshah/skillctl"
Repository = "https://github.com/dvlshah/skillctl"
Issues = "https://github.com/dvlshah/skillctl/issues"
Changelog = "https://github.com/dvlshah/skillctl/blob/main/CHANGELOG.md"

[project.scripts]
skillctl = "skillctl.cli:main"

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.40.0",
]
openai = [
    "openai>=2.0.0",
]
enrich = [
    "anthropic>=0.40.0",
    "openai>=2.0.0",
]
dev = [
    "pytest>=7.0",
    "anthropic>=0.40.0",
    "openai>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["skillctl"]
